Basic Realm persistence works Reorganized file structure
