+++ app/src/main/java/rocket/agile/com/mainlibrary2/MainActivity.java
@@ -1,5 +1,5 @@
 package rocket.agile.com.mainlibrary2;
 
-import rocket.agile.com.mainlibrary.MasterView;
+import rocket.agile.com.mainlibrary.activity.MasterView;
 
 public class MainActivity extends MasterView {}
