+++ main-library/src/main/AndroidManifest.xml
@@ -12,24 +12,24 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MasterView"
+            android:name=".activity.MasterView"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
             </intent-filter>
         </activity>
-        <activity android:name=".NavDrawerMain"
+        <activity android:name=".activity.NavDrawerMain"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity
-            android:name=".TabMenuMain"
+            android:name=".activity.TabMenuMain"
             android:label="@string/app_name"/>
         <activity
-            android:name=".GridButtonsActivity"
+            android:name=".activity.GridButtonsActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity
-            android:name=".LongButtonsActivity"
+            android:name=".activity.LongButtonsActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"/>
