+++ main-library/src/main/java/rocket/agile/com/mainlibrary/fragments/AboutUsFragment.java
@@ -1,15 +1,15 @@
-package rocket.agile.com.mainlibrary;
+package rocket.agile.com.mainlibrary.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import org.w3c.dom.Text;
+import rocket.agile.com.mainlibrary.R;
+import rocket.agile.com.mainlibrary.model.DataManager;
 
 
 /**
