+++ main-library/src/main/java/rocket/agile/com/mainlibrary/realm/RealmPersistence.java
@@ -1,10 +1,9 @@
-package rocket.agile.com.mainlibrary;
+package rocket.agile.com.mainlibrary.realm;
 
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
-import io.realm.RealmQuery;
-import io.realm.RealmResults;
-import rocket.agile.com.mainlibrary.ActionItems.LayoutValue;
+import rocket.agile.com.mainlibrary.actionItems.LayoutValue;
+import rocket.agile.com.mainlibrary.activity.MasterView;
 
 /**
  * Created by keithkowalski on 4/21/17.
@@ -29,16 +28,13 @@ public class RealmPersistence extends MasterView {
         Realm realm = Realm.getDefaultInstance();
         realm.beginTransaction();
 
-        try {
-            LayoutValue test = realm.where(LayoutValue.class).equalTo("id", LayoutValue.getId()).findFirst();
-            test.deleteFromRealm();
-        } catch (Exception e){
+        // Create Layout Value
+        LayoutValue layoutValue = new LayoutValue();
+        layoutValue.setId(0);
+        layoutValue.setLayoutValue(0);
 
-        }
-
-        LayoutValue managedLayoutValue = realm.createObject(LayoutValue.class);
-        managedLayoutValue.setId(0);
-        managedLayoutValue.setLayoutValue(1);
+        // Overwrite previous persisted object
+        realm.copyToRealmOrUpdate(layoutValue);
 
         realm.commitTransaction();
         realm.close();
