+++ main-library/src/main/res/layout/bottom_menu_activity_tab_menu_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="rocket.agile.com.mainlibrary.TabMenuMain">
+    tools:context="rocket.agile.com.mainlibrary.activity.TabMenuMain">
 
     <FrameLayout
         android:id="@+id/content"
