+++ main-library/src/main/res/layout/buttons_activity_long_buttons.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="rocket.agile.com.mainlibrary.LongButtonsActivity">
+    tools:context="rocket.agile.com.mainlibrary.activity.LongButtonsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
