+++ main-library/src/main/res/layout/content_grid_buttons.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="rocket.agile.com.mainlibrary.GridButtonsActivity"
+    tools:context="rocket.agile.com.mainlibrary.activity.GridButtonsActivity"
     tools:showIn="@layout/buttons_activity_grid_buttons">
 
 </android.support.v4.widget.NestedScrollView>
