+++ main-library/src/main/res/layout/fragment_about_us.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="rocket.agile.com.mainlibrary.AboutUsFragment">
+    tools:context="rocket.agile.com.mainlibrary.fragments.AboutUsFragment">
 
     <LinearLayout
         android:id="@+id/background_main"
