+++ main-library/src/main/res/layout/fragment_website.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="rocket.agile.com.mainlibrary.WebsiteFragment">
+    tools:context="rocket.agile.com.mainlibrary.fragments.WebsiteFragment">
 
     <WebView
         android:id="@+id/web_view"
