+++ main-library/src/main/res/layout/side_menu_content_nav_drawer_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="rocket.agile.com.mainlibrary.NavDrawerMain"
+    tools:context="rocket.agile.com.mainlibrary.activity.NavDrawerMain"
     tools:showIn="@layout/side_menu_app_bar_nav_drawer_main">
 
     <RelativeLayout
