+++ main-library/src/main/res/menu/menu_long_buttons.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="rocket.agile.com.mainlibrary.LongButtonsActivity">
+    tools:context="rocket.agile.com.mainlibrary.activity.LongButtonsActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
